FW: Fix a bunch of warnings in SAM3U build for CW-Lite/CW1200

+++ hardware/capture/chipwhisperer-lite/sam3u_fw/SAM3U_VendorExample/src/isp/V2Protocol.c
@@ -240,7 +240,7 @@ static void V2Protocol_LoadAddress(void)
 	if (CurrentAddress & (1UL << 31))
 	  MustLoadExtendedAddress = true;
 	  
-	printf("AVRPROG: Address = %02x\n", CurrentAddress);
+	printf("AVRPROG: Address = %02x\n", (unsigned int)CurrentAddress);
 
 	avrisp_status_payload[1] = STATUS_CMD_OK;
 	avrisp_status_payload_size = 2;
